Add initial version of btrfs-backup init script
My preference is to perform backups at specific events rather than in
certain intervals when they potentially interfer with other processes on
the system.
This change adds an init script that can be used to automatically run
btrfs-backup in order to backup the data of certain subvolumes. In its
current version, the init script is meant to be added to the shutdown
run level.