Disable the default timeout after which the script is killed
By default OpenRC imposes a timeout of 60s on every init script. That
means that if the script is not able to finish the backup within one
minute it will be forcibly killed (or some other measure will be taken,
I did not verify what happens).
In general, there is no upper bound on how long a backup can potentially
take because this depends largely on the amount being transferred as
well as the transfer speed. Nothing of this is predictable in a scenario
of synchronizing arbitrary btrfs subvolumes. To that end, this change
disables the timeout for the btrfs-backup script.