Enable automatic removal of old snapshots
For our frequently created snapshots we are not so much interested in
being able to jump back to an arbitrary snapshot created in the
potentially distant past. Rather, we mainly want to have the latest data
available on a backup in case something a disk crashes or the system
gets otherwise into a non-recoverable state.
This change adjust the init script to only keep snapshots for a week and
remove all older ones. The rationale is that if the system booted up
again a couple of times and the change was not reverted we assume
everything was fine and will forget about the previous state.