Enable debug mode
In the past we have seen one failure in btrfs-backup where not enough
information was available straight away in order to diagnose and debug
it properly.
This change passes the --debug switch to the btrfs-backup invocation
that that will cause it to report more information in case of an
exception.