Add support for custom error function
The btrfs-backup script has no real way of reporting errors in a manner
that is guaranteed to be visible to the user. In fact, in its current
style of being invoked during shutdown it is very unlikely that problems
are noticed at all because all error reporting happens on the screen and
just before the machine is rebooted (which in turn clears the screen).
There are many ways to get error reporting with more visibility,
however, the actual approach should be up to the user.
This change introduces a mechanism that allows the user to hook in a
custom error reporting function. By putting in place such a generic
mechanism the user is free to use any means that fit for reporting
errors.