Skip execution on reboots
In our use case, the system is not powered on 24/7 but only for part of
a day and shutdown subsequently. Hence, if a reboot is performed that is
being done because it has to be but it should not be seen as a moment
when a backup has to be performed, as that will happen once a shutdown
happens anyway.
This change skips performing a backup in case a reboot is performed.
Doing so speeds up reboots (potentially significantly) which improves
user experience.