Adjust init script to work in "normal" run levels
Until now the init script was designed to be executed in the 'shutdown'
run level. This approach causes problems with recent changes where we
might want to back up contents on an encrypted device (with the device
being managed by the dmcrypt run script). The problem is that all
'shutdown' scripts seem to execute only after those in the 'default' run
level are stopped, which is not so great.
This change reworks the script slightly to work in the "normal" run
levels (such as 'default'). Using it in the 'shutdown' level is not
supported anymore.