Report success in case of mount failure
The init script supports automatic mounting of devices to which to
backup. However, in its current design a failure doing so will cause the
entire init script to fail. Such a failure is not decided most of the
time because the script will remain in the "running" state.
This change treats mount failures (most of the time caused by a
[temporarily] missing removable device) as benign and simply exits the
script with an indication of success.