Add support for mounting both source and destination volume
So far the init script has supported cases where the destination volumes
volume was mounted on demand, but the source volume was always assumed
to be mounted already.
There are legitimate cases where the source volume would need to be
mounted on demand just as the destination volume. With this change we
add support for those use cases. To accomplish this task we split the
BTRFS_BACKUP_MOUNT variable into two, one representing the path to the
source volume, the other the path to the destination volume.